Reorganize MCP server
