+++ apps/mcp-server/src/mcp/tools/register-list-tables.ts
