+++ apps/mcp-server/src/mcp/tools/register-tools.ts
